Delete app_v2 | Refactor to app
