+++ .gitignore
@@ -8,7 +8,7 @@ local_settings.py
 db.sqlite3
 db.sqlite3-journal
 media
-app_v2/test.py
+app/test.py
 *.db
 .DS_Store
 node_modules
