+++ app/VANYA_check_this.txt
@@ -1,0 +1,21 @@
+Просто создал роутер, и парочку примочков для db
+и еще сделал так, чтобы при запуске первым делом запускалось init_db(app)
+там на 13 строчке app.on_event
+а так Tortoise прикольная штука, скажу честно удобнее чем sqlalchemy
+Еще я сделал хеширование пароля с помощью ебанной штучки из passlib
+Хз мне так удобнее, если хочешь там, можешь переписать мне ПОХ что будем использовать для
+хеширования пароля
+И да я еще все сделал app а не app_v2. Так будет удобнее.
+
+
+Все что ниже написал Copylot AHHAHHAHA Некоторые вещи думаю можно будет сделать.
+Я просто не знаю как сделать так, чтобы он мог писать в этот чатик.
+Еще кое-что я сделал, чтобы при запуске приложения, если в базе нет ни одного пользователя, то
+он создает админа с логином admin и паролем admin
+Так как что если ты запустишь приложение, то сразу сможешь зайти под админом.
+А потом уже будешь создавать новых пользователей и тд
+А еще я сделал так, чтобы при создании пользователя, он хешировал пароль.
+И еще я сделал так, чтобы при создании пользователя, он проверял, есть ли уже такой пользователь
+в базе, если есть, то он не создает его, а если нет, то создает
+
+Я сделал так, чтобы при регистрации пароль хешировался, а при авторизации пароль сравнивался с хешем
