--- app_v2/VANYA_check_this.txt
@@ -1,7 +1,0 @@
-Просто создал роутер, и парочку примочков для db
-и еще сделал так, чтобы при запуске первым делом запускалось init_db(app)
-там на 13 строчке app.on_event
-а так Tortoise прикольная штука, скажу честно удобнее чем sqlalchemy
-Еще я сделал хеширование пароля с помощью ебанной штучки из passlib
-Хз мне так удобнее, если хочешь там, можешь переписать мне ПОХ что будем использовать для
-хеширования пароля
