--- app_v2/models.py
@@ -1,24 +1,0 @@
-from tortoise import fields
-from tortoise.models import Model
-from tortoise.contrib.pydantic import pydantic_model_creator
-
-
-class Users(Model):
-    """
-    The User model
-    """
-
-    id = fields.IntField(pk=True, unique=True)
-    login = fields.CharField(max_length=50, null=False)
-    password = fields.CharField(max_length=128, null=False)
-    mcoin = fields.IntField(null=False)
-    rcoin = fields.IntField(null=False)
-
-
-# Pydantic schemas created automatically
-User_Pydantic = pydantic_model_creator(Users, name="User")
-UserIn_Pydantic = pydantic_model_creator(Users, name="UserIn", exclude_readonly=True)
-
-
-class Shop(Model):
-    pass
